# Improve internal function argument parsing performance by reducing code bloat