--- a/ext/gmp/gmp.c
+++ b/ext/gmp/gmp.c
@@ -137,15 +137,15 @@ static bool gmp_zend_parse_arg_into_mpz_ex(
 	 * but operator overloading with objects should behave as if a
 	 * method was called, thus strict types should apply. */
 	if (!ZEND_ARG_USES_STRICT_TYPES()) {
-		zend_long lval = 0;
 		if (is_operator && Z_TYPE_P(arg) == IS_NULL) {
 			return false;
 		}
-		if (!zend_parse_arg_long_weak(arg, &lval, arg_num)) {
+		zend_opt_long result = zend_parse_arg_long_weak(arg, arg_num);
+		if (!result.has_value) {
 			return false;
 		}
 
-		mpz_set_si(*destination_mpz_ptr, lval);
+		mpz_set_si(*destination_mpz_ptr, result.value);
 
 		return true;
 	}