--- a/ext/pgsql/pgsql.c
+++ b/ext/pgsql/pgsql.c
@@ -1900,7 +1900,7 @@ PHP_FUNCTION(pg_fetch_result)
 {
 	zval *result;
 	zend_string *field_name;
-	zend_long row, field_offset = 0;
+	zend_long row = 0, field_offset = 0;
 	bool row_is_null = false;
 	PGresult *pgsql_result;
 	pgsql_result_handle *pg_result;
@@ -2236,7 +2236,7 @@ static void php_pgsql_data_info(INTERNAL_FUNCTION_PARAMETERS, int entry_type, bo
 {
 	zval *result;
 	zend_string *field_name;
-	zend_long row, field_offset = 0;
+	zend_long row = 0, field_offset = 0;
 	bool row_is_null = false;
 	PGresult *pgsql_result;
 	pgsql_result_handle *pg_result;