… ext/standard/string.c