@nielsdos 2025-04-26
One potential optimization idea is to store a `bool is_empty` instead of `bool has_value`. Because in that case we have to store a 0 instead of a 1 for the "success path". This may be a bit faster on the x86 where setting the lower parts of some registers also clears the upper part.