@SakiTakamachi 2025-04-27
When building on Arm, an uninitialized warning was triggered at the following location.
Although there doesn’t seem to be any uninitialized code path, it might be a good idea to set an initial value just to silence the warning.

https://github.com/php/php-src/blob/c919ab4bdfaa472c95592437b1010bc284f55722/ext/date/php_date.c#L5514

-----

I also dumped the non-cold block assembly on Arm.
I already knew theoretically that it should have an effect on Arm as well, and it seems that is indeed the case.

before:
```
   0:   a9bc7bfd        stp     x29, x30, [sp, #-64]!
   4:   90000002        adrp    x2, 0 <__stack_chk_guard>
   8:   910003fd        mov     x29, sp
   c:   f9400042        ldr     x2, [x2]
  10:   a90153f3        stp     x19, x20, [sp, #16]
  14:   aa0103f4        mov     x20, x1
  18:   f90013f5        str     x21, [sp, #32]
  1c:   b9402c15        ldr     w21, [x0, #44]
  20:   f9400041        ldr     x1, [x2]
  24:   f9001fe1        str     x1, [sp, #56]
  28:   d2800001        mov     x1, #0x0                        // #0
  2c:   710006bf        cmp     w21, #0x1
  30:   54000401        b.ne    b0 <zif_twice+0xb0>  // b.any
  34:   39416001        ldrb    w1, [x0, #88]
  38:   7100103f        cmp     w1, #0x4
  3c:   54000221        b.ne    80 <zif_twice+0x80>  // b.any
  40:   f9402800        ldr     x0, [x0, #80]
  44:   d37ff800        lsl     x0, x0, #1
  48:   52800081        mov     w1, #0x4                        // #4
  4c:   f9000280        str     x0, [x20]
  50:   b9000a81        str     w1, [x20, #8]
  54:   90000000        adrp    x0, 0 <__stack_chk_guard>
  58:   f9400000        ldr     x0, [x0]
  5c:   f9401fe2        ldr     x2, [sp, #56]
  60:   f9400001        ldr     x1, [x0]
  64:   eb010042        subs    x2, x2, x1
  68:   d2800001        mov     x1, #0x0                        // #0
  6c:   540001c1        b.ne    a4 <zif_twice+0xa4>  // b.any
  70:   a94153f3        ldp     x19, x20, [sp, #16]
  74:   f94013f5        ldr     x21, [sp, #32]
  78:   a8c47bfd        ldp     x29, x30, [sp], #64
  7c:   d65f03c0        ret
  80:   91014013        add     x19, x0, #0x50
  84:   2a1503e2        mov     w2, w21
  88:   aa1303e0        mov     x0, x19
  8c:   9100c3e1        add     x1, sp, #0x30
  90:   94000000        bl      0 <zend_parse_arg_long_slow>
  94:   52800122        mov     w2, #0x9                        // #9
  98:   72001c1f        tst     w0, #0xff
  9c:   54000061        b.ne    a8 <zif_twice+0xa8>  // b.any
  a0:   14000000        b       0 <zif_twice>
  a4:   94000000        bl      0 <__stack_chk_fail>
  a8:   f9401be0        ldr     x0, [sp, #48]
  ac:   17ffffe6        b       44 <zif_twice+0x44>
  b0:   14000000        b       0 <zif_twice>
```

after:
```
   0:   a9bd7bfd        stp     x29, x30, [sp, #-48]!
   4:   910003fd        mov     x29, sp
   8:   f90013f5        str     x21, [sp, #32]
   c:   b9402c15        ldr     w21, [x0, #44]
  10:   a90153f3        stp     x19, x20, [sp, #16]
  14:   710006bf        cmp     w21, #0x1
  18:   540003a1        b.ne    8c <zif_twice+0x8c>  // b.any
  1c:   aa0103f4        mov     x20, x1
  20:   39416001        ldrb    w1, [x0, #88]
  24:   7100103f        cmp     w1, #0x4
  28:   54000141        b.ne    50 <zif_twice+0x50>  // b.any
  2c:   f9402800        ldr     x0, [x0, #80]
  30:   d37ff800        lsl     x0, x0, #1
  34:   52800081        mov     w1, #0x4                        // #4
  38:   f94013f5        ldr     x21, [sp, #32]
  3c:   f9000280        str     x0, [x20]
  40:   b9000a81        str     w1, [x20, #8]
  44:   a94153f3        ldp     x19, x20, [sp, #16]
  48:   a8c37bfd        ldp     x29, x30, [sp], #48
  4c:   d65f03c0        ret
  50:   91014013        add     x19, x0, #0x50
  54:   2a1503e1        mov     w1, w21
  58:   aa1303e0        mov     x0, x19
  5c:   94000000        bl      0 <zend_parse_arg_long_slow>
  60:   72001c3f        tst     w1, #0xff
  64:   54fffe61        b.ne    30 <zif_twice+0x30>  // b.any
  68:   52800120        mov     w0, #0x9                        // #9
  6c:   aa1303e4        mov     x4, x19
  70:   2a1503e1        mov     w1, w21
  74:   a94153f3        ldp     x19, x20, [sp, #16]
  78:   52800003        mov     w3, #0x0                        // #0
  7c:   f94013f5        ldr     x21, [sp, #32]
  80:   d2800002        mov     x2, #0x0                        // #0
  84:   a8c37bfd        ldp     x29, x30, [sp], #48
  88:   14000000        b       0 <zend_wrong_parameter_error>
  8c:   14000000        b       0 <zif_twice>
```

Regarding `is_empty`, the performance should theoretically be the same on Arm, so I think it’s fine to decide based on what’s best for x86.