@dstogov 2025-05-05
So, very light improvement on x86 Linux and degradation on Windows (both 32 and 64-bits). Correct?