@nielsdos 2025-05-05
> So, very light improvement on x86 Linux and degradation on Windows (both 32 and 64-bits). Correct?

That's correct. Degradation on 32-bit Windows is less severe than on 64-bit Windows.
I also tried wallclock time micro benchmarks but the execution time is too noisy to make any conclusions.